perf(Front/UMAP): Move UMAP Component logic to composables
