+++ front/src/composables/useColors.ts
@@ -1,11 +1,13 @@
 import {UMAPFiltersStore} from '../store/UMAP-filters.store';
-import type {Scale} from 'chroma-js';
+import type {Color, Scale} from 'chroma-js';
 import chroma from 'chroma-js';
 import type {ComputedRef} from 'vue';
 import {computed} from 'vue';
 
 interface UseColors {
     colors: ComputedRef<Scale>;
+    nightColor: Color;
+    dayColor: Color;
 }
 
 export function useColors(): UseColors {
@@ -20,7 +22,12 @@ export function useColors(): UseColors {
     throw new Error('Colors not defined');
   }
 
+  const nightColor = chroma('black');
+  const dayColor = colors.value(0);
+
   return {
     colors,
+    nightColor,
+    dayColor,
   };
 }
