+++ front/src/utils/is-hour-during-day.ts
@@ -1,0 +1,3 @@
+export function isHourDuringDay(hour: number): boolean {
+  return Math.abs(12 - hour) < 6;
+}
